+++ src/lib/pages/evm-contract-verify/components/EvmContractVerifyOptions.tsx
@@ -3,99 +3,120 @@ import { Control, useController, useWatch } from "react-hook-form";
 import {
   EvmContractVerifyForm,
   EvmProgrammingLanguage,
-  VerificationOptions,
+  VerifyOptions,
 } from "../types";
-import { getVerifyFormInitialValue } from "../helper";
 
 interface EvmContractVerifyVyperProps {
   control: Control<EvmContractVerifyForm>;
 }
+
+const EvmContractVerifyOptionsVyper = () => (
+  <>
+    <Radio
+      variant="gray-card"
+      width="fit-content"
+      value={VerifyOptions.VyperUploadFile}
+      overflow="hidden"
+      w="full"
+    >
+      Upload File
+    </Radio>
+    <Radio
+      variant="gray-card"
+      width="fit-content"
+      value={VerifyOptions.VyperContractCode}
+      overflow="hidden"
+      w="full"
+    >
+      Contract Code
+    </Radio>
+    <Radio
+      variant="gray-card"
+      width="fit-content"
+      value={VerifyOptions.VyperJsonInput}
+      overflow="hidden"
+      w="full"
+    >
+      JSON Input
+    </Radio>
+  </>
+);
+
+const EvmContractVerifyOptionsSolidity = () => (
+  <>
+    <Radio
+      variant="gray-card"
+      width="fit-content"
+      value={VerifyOptions.SolidityUploadFiles}
+      overflow="hidden"
+      w="full"
+    >
+      Upload File(s)
+    </Radio>
+    <Radio
+      variant="gray-card"
+      width="fit-content"
+      value={VerifyOptions.SolidityContractCode}
+      overflow="hidden"
+      w="full"
+    >
+      Contract Code
+    </Radio>
+    <Radio
+      variant="gray-card"
+      width="fit-content"
+      value={VerifyOptions.SolidityJsonInput}
+      overflow="hidden"
+      w="full"
+    >
+      JSON Input
+    </Radio>
+    <Radio
+      variant="gray-card"
+      width="fit-content"
+      value={VerifyOptions.SolidityHardhat}
+      overflow="hidden"
+      w="full"
+    >
+      Hardhat
+    </Radio>
+    <Radio
+      variant="gray-card"
+      width="fit-content"
+      value={VerifyOptions.SolidityFoundry}
+      overflow="hidden"
+      w="full"
+    >
+      Foundry
+    </Radio>
+  </>
+);
 
 export const EvmContractVerifyOptions = ({
   control,
 }: EvmContractVerifyVyperProps) => {
   const { field } = useController({
     control,
-    name: "verifyForm.form",
+    name: "option",
   });
 
-  const [language, verifyFormOption] = useWatch({
+  const language = useWatch({
     control,
-    name: ["verifyForm.language", "verifyForm.form.option"],
+    name: "language",
   });
-
-  const handleOptionChange = (nextVal: VerificationOptions) => {
-    field.onChange(getVerifyFormInitialValue(language, nextVal));
-  };
 
   return (
     <Stack spacing={6}>
       <Heading as="h6" variant="h6">
         Select Verification Option
       </Heading>
-      <RadioGroup onChange={handleOptionChange} value={verifyFormOption}>
+      <RadioGroup onChange={(val) => field.onChange(val)} value={field.value}>
         <Grid gridTemplateColumns="repeat(3, 1fr)" gap={4}>
-          {language === EvmProgrammingLanguage.Solidity && (
-            <Radio
-              variant="gray-card"
-              width="fit-content"
-              value={VerificationOptions.UploadFiles}
-              overflow="hidden"
-              w="full"
-            >
-              Upload File(s)
-            </Radio>
-          )}
           {language === EvmProgrammingLanguage.Vyper && (
-            <Radio
-              variant="gray-card"
-              width="fit-content"
-              value={VerificationOptions.UploadFile}
-              overflow="hidden"
-              w="full"
-            >
-              Upload File
-            </Radio>
+            <EvmContractVerifyOptionsVyper />
           )}
-          <Radio
-            variant="gray-card"
-            width="fit-content"
-            value={VerificationOptions.ContractCode}
-            overflow="hidden"
-            w="full"
-          >
-            Contract Code
-          </Radio>
-          <Radio
-            variant="gray-card"
-            width="fit-content"
-            value={VerificationOptions.JsonInput}
-            overflow="hidden"
-            w="full"
-          >
-            JSON Input
-          </Radio>
           {language === EvmProgrammingLanguage.Solidity && (
-            <>
-              <Radio
-                variant="gray-card"
-                width="fit-content"
-                value={VerificationOptions.Hardhat}
-                overflow="hidden"
-                w="full"
-              >
-                Hardhat
-              </Radio>
-              <Radio
-                variant="gray-card"
-                width="fit-content"
-                value={VerificationOptions.Foundry}
-                overflow="hidden"
-                w="full"
-              >
-                Foundry
-              </Radio>
-            </>
+            <EvmContractVerifyOptionsSolidity />
           )}
         </Grid>
       </RadioGroup>